requirements-dev.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile requirements-dev.ini
#
annotated-types==0.7.0
    # via pydantic
anyio==4.11.0
    # via
    #   google-genai
    #   httpx
black==25.11.0
    # via -r requirements-dev.ini
cachetools==6.2.2
    # via google-auth
certifi==2025.11.12
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.4
    # via requests
click==8.3.1
    # via black
colorama==0.4.6
    # via click
flake8==7.0.0
    # via pyproject-flake8
google-auth==2.43.0
    # via google-genai
google-genai==1.52.0
    # via -r requirements.ini
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via google-genai
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
isort==7.0.0
    # via -r requirements-dev.ini
mccabe==0.7.0
    # via flake8
mypy==1.18.2
    # via -r requirements-dev.ini
mypy-extensions==1.1.0
    # via
    #   black
    #   mypy
packaging==25.0
    # via black
pathspec==0.12.1
    # via
    #   black
    #   mypy
platformdirs==4.5.0
    # via black
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pycodestyle==2.11.1
    # via flake8
pydantic==2.12.5
    # via google-genai
pydantic-core==2.41.5
    # via pydantic
pyflakes==3.2.0
    # via flake8
pyproject-flake8==7.0.0
    # via -r requirements-dev.ini
python-dotenv==1.2.1
    # via -r requirements.ini
pytokens==0.3.0
    # via black
pyyaml==6.0.3
    # via -r requirements.ini
requests==2.32.5
    # via google-genai
rsa==4.9.1
    # via google-auth
sniffio==1.3.1
    # via anyio
tenacity==9.1.2
    # via google-genai
typing-extensions==4.15.0
    # via
    #   google-genai
    #   mypy
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
urllib3==2.5.0
    # via requests
websockets==15.0.1
    # via google-genai
